Refatorando controlador

+++ src/main/java/br/com/syscam/rest/AdminController.java
@@ -22,19 +22,19 @@ public class AdminController {
     public @ResponseBody boolean salvar(@RequestBody Administrador administrador) {
 
         try {
-            if(!buscar(administrador.getCodigo()).isPresent()){
-                repository.save(administrador);
-                return true;
-            }
-            return false;
-
+           if(buscar(administrador.getCodigo()).isPresent()){
+               return false;
+           }else {
+               repository.save(administrador);
+               return true;
+           }
         } catch (Exception e) {
 
             return false;
         }
     }
 
-    @GetMapping("{codigo}")
+    @GetMapping("/{codigo}")
     public @ResponseBody Optional<Administrador> buscar(@PathVariable("codigo") Integer codigo) {
         return repository.findById(codigo);
     }
@@ -47,7 +47,7 @@ public class AdminController {
 
     @PutMapping
     public @ResponseBody boolean atualizar(@RequestBody Administrador administrador) {
-        if (buscar((administrador.getCodigo())).isPresent()) {
+        if (buscar(administrador.getCodigo()).isPresent()) {
             this.repository.save(administrador);
             return true;
         }
@@ -55,15 +55,13 @@ public class AdminController {
     }
 
     @DeleteMapping("/{codigo}")
-    public @ResponseBody boolean remover(@RequestBody Integer codigo){
+    public @ResponseBody boolean remover(@PathVariable("codigo") Integer codigo){
         try{
-//            if(buscar(email).isPresent()){
-//                repository.deleteById(email);
-//                return true;
-//            }
-//            return false;
-            repository.deleteById(codigo);
-            return true;
+            if(buscar(codigo).isPresent()){
+                repository.deleteById(codigo);
+                return true;
+            }
+            return false;
         }catch (Exception e){
             return false;
         }
